requirements.txt
textual
ollama
ddgs
beautifulsoup4
lxml
sqlite-utils
sentence-transformers
faiss-cpu
numpy
fastapi
uvicorn
websockets
httpx
flask>=2.3.0
requests>=2.28.0
firebase>=4.0.1
python-dotenv>=1.0.0
pyrebase4>=4.7.1
pyyaml>=6.0
firebase-admin>=6.4.0
google-cloud-firestore>=2.13.0
